requirements.txt
garminconnect>=0.2.0
python-telegram-bot>=20.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
tenacity>=8.2.0
matplotlib>=3.7.0
anthropic>=0.39.0
pyzbar>=0.1.9
Pillow>=10.0.0
requests>=2.31.0
